Cleaned up label import code.
