+++ extensions/dicom-sr/src/utils/getLabelFromDCMJSImportedToolData.js
@@ -1,0 +1,17 @@
+export default function getLabelFromDCMJSImportedToolData(toolData) {
+  const { findingSites = [], findings = [] } = toolData;
+
+  let freeTextLabel = findingSites.find(
+    fs => fs.CodeValue === 'CORNERSTONEFREETEXT'
+  );
+
+  if (freeTextLabel) {
+    return freeTextLabel.CodeMeaning;
+  }
+
+  freeTextLabel = findings.find(f => f.CodeValue === 'CORNERSTONEFREETEXT');
+
+  if (freeTextLabel) {
+    return freeTextLabel.CodeMeaning;
+  }
+}
