+++ extensions/dicom-sr/src/utils/getToolStateToCornerstoneMeasurementSchema.js
@@ -51,7 +51,7 @@ function Length(measurementData, imageId, _getValueTypeFromToolType) {
     StudyInstanceUID,
   } = instance;
 
-  const { handles, findingSites, findings } = measurementData;
+  const { handles, label } = measurementData;
 
   const points = [];
   Object.keys(handles).map(handle => {
@@ -75,24 +75,8 @@ function Length(measurementData, imageId, _getValueTypeFromToolType) {
     length: measurementData.length,
     type: _getValueTypeFromToolType(tool),
     points,
-    label: _getLabel(findingSites, findings),
+    label,
   };
-}
-
-function _getLabel(findingSites = [], findings = []) {
-  let freeTextLabel = findingSites.find(
-    fs => fs.CodeValue === 'CORNERSTONEFREETEXT'
-  );
-
-  if (freeTextLabel) {
-    return freeTextLabel.CodeMeaning;
-  }
-
-  freeTextLabel = findings.find(f => f.CodeValue === 'CORNERSTONEFREETEXT');
-
-  if (freeTextLabel) {
-    return freeTextLabel.CodeMeaning;
-  }
 }
 
 function Bidirectional(measurementData, imageId, _getValueTypeFromToolType) {
@@ -105,7 +89,7 @@ function Bidirectional(measurementData, imageId, _getValueTypeFromToolType) {
     StudyInstanceUID,
   } = instance;
 
-  const { handles, findingSites, findings } = measurementData;
+  const { handles, label } = measurementData;
 
   const longAxis = [handles.start, handles.end];
   const shortAxis = [handles.perpendicularStart, handles.perpendicularEnd];
@@ -123,7 +107,7 @@ function Bidirectional(measurementData, imageId, _getValueTypeFromToolType) {
     longestDiameter: measurementData.longestDiameter,
     type: _getValueTypeFromToolType(tool),
     points: { longAxis, shortAxis },
-    label: _getLabel(findingSites, findings),
+    label,
   };
 }
 
@@ -137,7 +121,9 @@ function EllipticalRoi(measurementData, imageId, _getValueTypeFromToolType) {
     StudyInstanceUID,
   } = instance;
 
-  const { handles, findingSites, findings } = measurementData;
+  const { handles, label } = measurementData;
+
+  debugger;
 
   const { start, end } = handles;
 
@@ -181,7 +167,7 @@ function EllipticalRoi(measurementData, imageId, _getValueTypeFromToolType) {
         .area /* TODO: Add concept names instead (descriptor) */,
     type: _getValueTypeFromToolType(tool),
     points,
-    label: _getLabel(findingSites, findings),
+    label,
   };
 }
 
@@ -195,7 +181,7 @@ function ArrowAnnotate(measurementData, imageId, _getValueTypeFromToolType) {
     StudyInstanceUID,
   } = instance;
 
-  const { handles, findingSites, findings } = measurementData;
+  const { handles } = measurementData;
 
   const points = [];
   Object.keys(handles).map(handle => {
@@ -219,6 +205,6 @@ function ArrowAnnotate(measurementData, imageId, _getValueTypeFromToolType) {
     text: measurementData.text,
     type: _getValueTypeFromToolType(tool),
     points,
-    label: _getLabel(findingSites, findings),
+    label,
   };
 }
